Finished GuildRelated stuff except like Dispatch Commanders for gold and Conquest related

+++ StellarGK/Database/Schemes/GameProfileScheme.cs
@@ -16,7 +16,7 @@ namespace StellarGK.Database.Schemes
         public int LastStage { get; set; }
         public bool Notifaction { get; set; }
         public int ResetDateTime { get; set; }
-        public int LastLoginTime { get; set; }
+        public double LastLoginTime { get; set; }
         public UserDevice UserDevice { get; set; }
         public UserBattleStatistics UserStatistics { get; set; }
         public UserResources UserResources { get; set; }
